# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "REST API for Topcoder Submissions.\n\n### Pagination\n  Requests that\
    \ return multiple items will be paginated to 20 items by default. You can specify\n\
    \  further pages with the `page` parameter. You can also set a custom page\n \
    \ size up to 100 with the `perPage` parameter.\n\n  Pagination response data is\
    \ included in http headers. By Default, the response header contains links with\
    \ `next`, `last`, `first`, `prev` resource links.\n"
  version: "5.0.0"
  title: "Topcoder Submissions REST API"
  termsOfService: "http://topcoder.com/terms/"
  contact:
    url: "http://www.topcoder.com/support"
    email: "support@topcoder.com"
host: "api.topcoder.com"
basePath: "/v5"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security:
- Bearer: []
paths:
  /submissions:
    get:
      tags:
      - "Submissions"
      summary: "Get all submissions."
      description: "Get all submissions. Link headers are sent back and they have\
        \ rel set to prev, next, first, last and contain the relevant URL.\n\nResulted\
        \ collection of submissions can be filtered using filter parameters `type`,\
        \ `url`, `memberId`, `challengeId` (all filter parameters are optional and\
        \ combined by the logical operation `AND`).\n\n**Authorization:** Submission\
        \ is accessible by roles `topcoder user`, `admin` and `copilot`.\n"
      operationId: "getSubmissions"
      parameters:
      - name: "page"
        in: "query"
        description: "The page number."
        required: false
        type: "integer"
        default: 1
      - name: "perPage"
        in: "query"
        description: "The number of items to list per page."
        required: false
        type: "integer"
        default: 20
        maximum: 100
      - name: "type"
        in: "query"
        description: "The type filter for submissions."
        required: false
        type: "string"
      - name: "url"
        in: "query"
        description: "The url filter for submissions."
        required: false
        type: "string"
      - name: "memberId"
        in: "query"
        description: "The member id filter for submissions."
        required: false
        type: "string"
      - name: "challengeId"
        in: "query"
        description: "The challenge id filter for submissions."
        required: false
        type: "string"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - the request was successful, the submissions resource\
            \ are returned."
          schema:
            type: "array"
            example:
            - id: "f11668564ca74c265de7009b"
              type: "ContestSubmission"
              url: "https://software.topcoder.com/review/actions/DownloadContestSubmission?uid=123456"
              memberId: "d40796a325341bdb8c973656"
              challengeId: "8b1a70e7e492adc4551a0d68"
              created: "2018-05-20T07:00:30.123Z"
              updated: "2018-06-01T07:36:28.178Z"
              createdBy: "topcoder user"
              updatedBy: "topcoder user"
            - id: "c160c47299589a2d90bd2247"
              type: "ContestSubmission"
              url: "https://software.topcoder.com/review/actions/DownloadContestSubmission?uid=123457"
              memberId: "d40796a325341bdb8c973656"
              challengeId: "8b1a70e7e492adc4551a0d68"
              created: "2018-05-20T08:00:30.000Z"
              updated: "2018-06-01T09:23:00.178Z"
              createdBy: "topcoder user"
              updatedBy: "topcoder user"
            items:
              $ref: "#/definitions/Submission"
          headers:
            X-Next-Page:
              type: "integer"
              description: "The index of the next page"
            X-Page:
              type: "integer"
              description: "The index of the current page (starting at 1)"
            X-Per-Page:
              type: "integer"
              description: "The number of items to list per page"
            X-Prev-Page:
              type: "integer"
              description: "The index of the previous page"
            X-Total:
              type: "integer"
              description: "The total number of items"
            X-Total-Pages:
              type: "integer"
              description: "The total number of pages"
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
            Link:
              type: "string"
              description: "Pagination link header."
        304:
          $ref: "#/definitions/NotModified"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Submissions"
    head:
      tags:
      - "Submissions"
      summary: "Get only response status and headers information but no response body\
        \ for the endpoint.\n"
      description: "Get response status and headers information for the endpoint.\
        \ Link headers are sent back and they have rel set to prev, next, first, last\
        \ and contain the relevant URL.\n\nRequested submissions can be filtered using\
        \ filter parameters `type`, `url`, `memberId`, `challengeId` (all filter parameters\
        \ are optional and combined by the logical operation `AND`).\n\n**Authorization:**\
        \ Submission is accessible by roles `topcoder user`, `admin` and `copilot`.\n"
      operationId: "headSubmissions"
      parameters:
      - name: "page"
        in: "query"
        description: "The page number."
        required: false
        type: "integer"
        default: 1
      - name: "perPage"
        in: "query"
        description: "The number of items to list per page."
        required: false
        type: "integer"
        default: 20
        maximum: 100
      - name: "type"
        in: "query"
        description: "The type filter for submissions."
        required: false
        type: "string"
      - name: "url"
        in: "query"
        description: "The url filter for submissions."
        required: false
        type: "string"
      - name: "memberId"
        in: "query"
        description: "The member id filter for submissions."
        required: false
        type: "string"
      - name: "challengeId"
        in: "query"
        description: "The challenge id filter for submissions."
        required: false
        type: "string"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - The request was successful."
          headers:
            X-Next-Page:
              type: "integer"
              description: "The index of the next page"
            X-Page:
              type: "integer"
              description: "The index of the current page (starting at 1)"
            X-Per-Page:
              type: "integer"
              description: "The number of items to list per page"
            X-Prev-Page:
              type: "integer"
              description: "The index of the previous page"
            X-Total:
              type: "integer"
              description: "The total number of items"
            X-Total-Pages:
              type: "integer"
              description: "The total number of pages"
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
            Link:
              type: "string"
              description: "Pagination link header."
        304:
          $ref: "#/definitions/NotModifiedWithoutBody"
        400:
          $ref: "#/definitions/BadRequestWithoutBody"
        401:
          $ref: "#/definitions/UnauthorizedWithoutBody"
        403:
          $ref: "#/definitions/ForbiddenWithoutBody"
        500:
          $ref: "#/definitions/ServerErrorWithoutBody"
      x-swagger-router-controller: "Submissions"
    post:
      tags:
      - "Submissions"
      summary: "Create a submission."
      description: "Create a new submission.\n\n**Authorization:** Submission creation\
        \ is accessible by roles `topcoder user`, `admin` and `copilot`.\n"
      operationId: "createSubmission"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Submission"
      responses:
        201:
          description: "Created - The request was successful and the resource is returned."
          schema:
            $ref: "#/definitions/Submission"
          headers:
            Location:
              type: "string"
              description: "The Location response header indicates the URL to redirect\
                \ a page to or created resource. It only provides a meaning when served\
                \ with a 3xx (redirection) or 201 (created) status response."
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        409:
          $ref: "#/definitions/Conflict"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Submissions"
  /submissions/{submissionId}:
    get:
      tags:
      - "Submissions"
      summary: "Get the submission."
      description: "Get the submission by id.\n\n**Authorization:** Submission is\
        \ accessible by roles `topcoder user`, `admin` and `copilot`.\n"
      operationId: "getSubmission"
      parameters:
      - name: "submissionId"
        in: "path"
        description: "submission id"
        required: true
        type: "string"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - the request was successful and the submission is returned."
          schema:
            $ref: "#/definitions/Submission"
          headers:
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
        304:
          $ref: "#/definitions/NotModified"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Submissions"
    head:
      tags:
      - "Submissions"
      summary: "Get only response status and headers information but no response body\
        \ for the endpoint.\n"
      description: "Get response status and headers information for the endpoint.\
        \ It does not contain response body.\n\n**Authorization:** Submission is accessible\
        \ by roles `topcoder user`, `admin` and `copilot`.\n"
      operationId: "headSubmission"
      parameters:
      - name: "submissionId"
        in: "path"
        description: "submission id"
        required: true
        type: "string"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - The request was successful."
          headers:
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
        304:
          $ref: "#/definitions/NotModifiedWithoutBody"
        400:
          $ref: "#/definitions/BadRequestWithoutBody"
        401:
          $ref: "#/definitions/UnauthorizedWithoutBody"
        403:
          $ref: "#/definitions/ForbiddenWithoutBody"
        500:
          $ref: "#/definitions/ServerErrorWithoutBody"
      x-swagger-router-controller: "Submissions"
    put:
      tags:
      - "Submissions"
      summary: "Update the submission."
      description: "Update the submission by id.\n\n**Authorization:** Submission\
        \ update is accessible only by `admin` role.\n"
      operationId: "updateSubmission"
      parameters:
      - name: "submissionId"
        in: "path"
        description: "submission id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/UpdatableSubmission"
      responses:
        200:
          description: "OK - The request was successful and the updated resource is\
            \ returned."
          schema:
            $ref: "#/definitions/Submission"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Submissions"
    delete:
      tags:
      - "Submissions"
      summary: "Delete the submission."
      description: "Delete the submission.\n\n**Authorization:** Submission deletion\
        \ is accessible only by `admin` role.\n"
      operationId: "deleteSubmission"
      parameters:
      - name: "submissionId"
        in: "path"
        description: "submission id"
        required: true
        type: "string"
      responses:
        204:
          $ref: "#/definitions/NoContent"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Submissions"
    patch:
      tags:
      - "Submissions"
      summary: "Partially update the submission."
      description: "Allows to partially modify the submission with the provided request\
        \ body properties.\n\n**Authorization:** Partially modify of submission is\
        \ accessible only by `admin` role.\n"
      operationId: "partiallyUpdateSubmission"
      parameters:
      - name: "submissionId"
        in: "path"
        description: "submission id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/PartiallySubmission"
      responses:
        200:
          description: "OK - The request was successful and the updated resource is\
            \ returned."
          schema:
            $ref: "#/definitions/Submission"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "Submissions"
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  NoContent:
    description: "No Content - The server has successfully fulfilled the request and\
      \ that there is no additional content to send in the response payload body.\n"
  BadRequestWithoutBody:
    description: "Bad Request - Request validation error, unable to process the request\
      \ due to an error (e.g., malformed request syntax, size too large, invalid request\
      \ message framing, or deceptive request routing).\n"
  UnauthorizedWithoutBody:
    description: "Unauthorized - Occurs when the JWT token provided is either expired\
      \ or invalid.\n"
  ForbiddenWithoutBody:
    description: "Forbidden - The request is not allowed.\n"
  NotModifiedWithoutBody:
    description: "Not Modified - Indicates that the resource has not been modified\
      \ since the last request.\n"
  ServerErrorWithoutBody:
    description: "Server Error - While handling the request something went wrong server-side.\n"
  Submission:
    allOf:
    - type: "object"
      properties:
        id:
          type: "string"
          example: "f11668564ca74c265de7009b"
          description: "The submission id."
          readOnly: true
    - $ref: "#/definitions/PartiallySubmission"
    - $ref: "#/definitions/Auditable"
    description: "The submission entity."
  PartiallySubmission:
    type: "object"
    properties:
      type:
        type: "string"
        example: "ContestSubmission"
        description: "The submission type."
      url:
        type: "string"
        example: "http://content.topcoder.com/some/path"
        description: "The submission url."
      memberId:
        type: "string"
        example: "d40796a325341bdb8c973656"
        description: "The submission member id."
      challengeId:
        type: "string"
        example: "8b1a70e7e492adc4551a0d68"
        description: "The submission challenge id."
    description: "The submission entity fields that can be partially updated."
    example:
      challengeId: "8b1a70e7e492adc4551a0d68"
      type: "ContestSubmission"
      url: "http://content.topcoder.com/some/path"
      memberId: "d40796a325341bdb8c973656"
  UpdatableSubmission:
    allOf:
    - $ref: "#/definitions/PartiallySubmission"
    description: "The submission entity fields that updates whole entity."
...
  Auditable:
    type: "object"
    properties:
      created:
        type: "string"
        format: "date-time"
        example: "2018-05-20T07:00:30.123Z"
        description: "The entity creation date."
        readOnly: true
      updated:
        type: "string"
        format: "date-time"
        example: "2018-06-01T07:36:28.178Z"
        description: "The entity update date."
        readOnly: true
      createdBy:
        type: "string"
        example: "admin"
        description: "The user who created the entity."
        readOnly: true
      updatedBy:
        type: "string"
        example: "admin"
        description: "The user who updated the entity."
        readOnly: true
    description: "The entity used as mark to add audit fields."
  Unauthorized:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Unable to authenticate the user."
        description: "The unauthorized error message."
    description: "The unauthorized error entity."
  NotFound:
    type: "object"
    properties:
      message:
        type: "string"
        example: "A resource with the name could not be found."
        description: "The not found error message."
    description: "The not found error entity."
  NotModified:
    type: "object"
    properties:
      message:
        type: "string"
        example: "There is no need to transfer a representation of the target resource\
          \ because the request indicates that the client, which made the request\
          \ conditional, already has a valid representation."
        description: "The not modified error message."
    description: "The not modified entity."
  ServerError:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Something went wrong while processing your request. We’re sorry\
          \ for the trouble. We’ve been notified of the error and will correct it\
          \ as soon as possible. Please try your request again in a moment."
        description: "The server error message."
    description: "The server error entity."
  BadRequest:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Invalid input."
        description: "The bad request error message."
    description: "The bad request error entity."
  Forbidden:
    type: "object"
    properties:
      message:
        type: "string"
        example: "You are not allowed to access the request."
        description: "The forbidden error message."
    description: "The permission error entity."
  Conflict:
    type: "object"
    required:
    - "message"
    properties:
      message:
        type: "string"
        example: "Creating a resource with a name already exists."
        description: "The conflict error message."
    description: "The conflict error entity."
...